Change data type

+++ GPU_CUDA_Distances/kernel.cu
@@ -45,22 +45,21 @@ __global__ void DotProduct
 
 
 __global__ void FirstMerge
-(int N, embed_t *sims,unsigned int* pos, int length, int pad) {
-
+(int N, embed_t *sims,unsigned int* pos, int64_t length, int64_t pad) {
   
-    int id = blockIdx.x * blockDim.x + threadIdx.x;
-    int start=id*N;
-    int end=start+N;
+    int64_t id = blockIdx.x * blockDim.x + threadIdx.x;
+    int64_t start=id*N;
+    int64_t end=start+N;
     if (start<length) { 
     
     // Insertion sort, as N SHOULD be small
     
-	for(int i=start+1; i<end; i++)
+	for(int64_t i=start+1; i<end; i++)
 	{
     if (i<length){
 		embed_t temp=sims[i];
-        unsigned int position=pos[i];
-		int j=i-1;
+        int64_t position=pos[i];
+		int64_t j=i-1;
 		while((temp>sims[j]) && (j>=start))
 		{
 			sims[j+1]=sims[j];
@@ -71,7 +70,7 @@ __global__ void FirstMerge
         pos[(j+1)]=position;
 	}
     else if (i<pad) {
-        for (int i=0;i<N;++i) {
+        for (int64_t i=0;i<N;++i) {
     	sims[id+i]=0;
         pos[id+i]=0;
         }
